Split JEI.java put private classes in their own package
